Update patch set 1
Patch Set 1: -Code-Review

Honestly, I don't know what we should do with CAF strings. We already merged a change to cleanup the strings in BluetoothExt in the past and this change is following the same logic.

Generally CAF strings quality is bad, but since we can't mod the strings, we are stuck with bad English strings everywhere when we merge from CAF. Maybe we're overreacting about touching these strings and should allow modifying?

Of course this might create merge conflicts, but is avoiding merge conflicts worth the loss of good quality strings?

Patch-set: 1